Add clang-tidy checking to build
make_rules/arm_gcc_rules.mak:
* Separate ARM_FLAGS from CXX_FLAGS, C_FLAGS, AS_FLAGS, LD_FLAGS
  in gcc compiler tool chain.
* Add clang tidy command to compiler command.

make_rules/checker_rules.mak:
  Add clang-tidy definitions and rules.
  Output errors and warning to _checks/<source_file_name>.<ext>.check
To invoke checking: make checks

Use angle-brackets for Nordic SDK header files,
  including softdevice headers.
  Makefile: Use Nordic system files
  Change gcc include directive from -I to -isystem

project_assert.h:
  Use stdlib assert() when not compiling for ARM.
    This eliminates a raft of tidy conditional complaints.
  Rename unit_tests/itest_assert.cc -> assert_stubs.cc
    This file is still compiled in, yet no longer used.
    It hurts nothing, so keep it in case it is useful later.

+++ nordic/peripherals/spis.cc
@@ -340,10 +340,10 @@ void spis_deinit(spi_port_t spi_port)
 
 static void spis_arm_transfer(struct spis_control_block_t* spis_control)
 {
-    spis_control->spis_registers->TXD.PTR    = reinterpret_cast<uint32_t>(spis_control->tx_buffer);
+    spis_control->spis_registers->TXD.PTR    = reinterpret_cast<uintptr_t>(spis_control->tx_buffer);
     spis_control->spis_registers->TXD.MAXCNT = spis_control->tx_length;
 
-    spis_control->spis_registers->RXD.PTR    = reinterpret_cast<uint32_t>(spis_control->rx_buffer);
+    spis_control->spis_registers->RXD.PTR    = reinterpret_cast<uintptr_t>(spis_control->rx_buffer);
     spis_control->spis_registers->RXD.MAXCNT = spis_control->rx_length;
 
     // Release the SPI slave semaphore from CPU ownership.
